Button for new added on the JournalPage aswell as refactoring the AppRouter page
