+++ src/hooks/useCheckAuth.js
@@ -1,0 +1,20 @@
+import { useSelector, useDispatch } from "react-redux";
+import { onAuthStateChanged } from "firebase/auth";
+import { FirebaseAuth } from "../firebase/config";
+import { useEffect } from "react";
+import { logout, login } from "../store/auth/authSlice";
+
+export const useCheckAuth = () => {
+  const { status } = useSelector((state) => state.auth);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    onAuthStateChanged(FirebaseAuth, async (user) => {
+      if (!user) return dispatch(logout());
+      const { uid, email, displayName, photoURL } = user;
+      dispatch(login({ uid, email, displayName, photoURL }));
+    });
+  }, []);
+
+  return { status };
+};
