+++ src/router/AppRouter.jsx
@@ -1,24 +1,11 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 import AuthRoutes from "../auth/routes/AuthRoutes";
 import JournalRoutes from "../journal/routes/JournalRoutes";
-import { useSelector, useDispatch } from "react-redux";
 import CheckingAuth from "../ui/components/CheckingAuth";
-import { onAuthStateChanged } from "firebase/auth";
-import { FirebaseAuth } from "../firebase/config";
-import { useEffect } from "react";
-import { logout, login } from "../store/auth/authSlice";
+import { useCheckAuth } from "../hooks/useCheckAuth";
 
 export const AppRouter = () => {
-  const { status } = useSelector((state) => state.auth);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    onAuthStateChanged(FirebaseAuth, async (user) => {
-      if (!user) return dispatch(logout());
-      const { uid, email, displayName, photoURL } = user;
-      dispatch(login({ uid, email, displayName, photoURL }));
-    });
-  }, []);
+  const { status } = useCheckAuth();
 
   if (status === "checking") {
     return <CheckingAuth />;
